[controller] Extract StoreRecoveryExecutor from DegradedModeRecoveryService
Splits DegradedModeRecoveryService (511 LOC, over the 500-line lint
threshold) into two cohesive classes. No behavior change.

DegradedModeRecoveryService (305 LOC) keeps cross-store orchestration:
- triggerRecovery (Phase 1: prepare + initiate per store)
- confirmRecoveryAndTransitionVersions (Phase 2: poll completion + transition)
- findPartiallyOnlineStores (discovery)
- thread pools, progress tracking, lifecycle, public API

StoreRecoveryExecutor (251 LOC) owns per-store recovery operations:
- recoverSingleStore (pre-check + retry/backoff around
  prepare → poll-ready → initiate)
- pollUntilReady (readiness polling helper)
- resolveSourceFabric (source-fabric resolution helper)
- pollUntilVersionCurrent + VersionPollResult enum (Phase 2 polling)
- Owns the constants and tunables relevant to per-store recovery:
  MAX_RETRIES, READINESS_POLL_*, DEFAULT_RECOVERY_COMPLETION_POLL_*.

setRecoveryCompletionPollParameters on the service forwards to the
executor so existing tests don't need to learn about the executor.
